feat(get-executed-tasks): add in memory db to store and get executed tasks

+++ CodeMasters.Domain/Aggregates/Operations/Division.cs
@@ -3,7 +3,7 @@
     public class Division : IOperation
     {
 
-        public decimal Calculate(decimal left, decimal right)
+        public double Calculate(double left, double right)
         {
             if (right == 0) throw new DivideByZeroException();
             return left / right;
