+++ CodeMasters.Domain/Aggregates/Operations/Subtraction.cs
@@ -2,7 +2,7 @@
 {
     public class Subtraction : IOperation
     {
-        public decimal Calculate(decimal left, decimal right) => left - right;
+        public double Calculate(double left, double right) => left - right;
 
     }
 }
